Splitting of score_engine.py file to simplify
This code (split into scoring_main.py, scoring_logic.py, data_fetch.py, and config.py) should function exactly the same as the score_engine.py. I also included a .gitignore for the __pycache__ directory that it creates when it runs since that is dependent on the platform it is running on.